requirements.txt
fhirpy==1.3.0
fhir.resources==6.4.0
iris_pex_embedded_python
https://github.com/grongierisc/iris-embedded-python-wrapper/releases/download/v0.0.3/iris-0.0.3-py3-none-any.whl